+++ Arimaa/src/main/java/cz/cvut/fel/pjv/arimaa/ui/JavafxApp.java
@@ -1,14 +1,25 @@
 package cz.cvut.fel.pjv.arimaa.ui;
 
+import cz.cvut.fel.pjv.arimaa.controller.GameController;
+import cz.cvut.fel.pjv.arimaa.model.Game;
 import javafx.application.Application;
 import javafx.stage.Stage;
 
 /**
- * JavaFX application entry; scene graph and controllers will be added later.
+ * JavaFX entry: wires {@link Game} / {@link GameController} / {@link MainController} and shows setup UI.
  */
 public class JavafxApp extends Application {
 
     @Override
     public void start(Stage primaryStage) {
+        Game game = new Game();
+        game.startNewGame();
+
+        GameController gameController = new GameController();
+        gameController.setGame(game);
+
+        MainController mainController = new MainController();
+        mainController.setGameController(gameController);
+        mainController.attachToStage(primaryStage);
     }
 }